[Refactor] 클래스 위치 및 패키지 구조 리팩토링 (#150)
* refactor: 클래스 위치 및 패키지 구조 리팩토링

* test: 리팩토링에 따른 test 수정
