+++ Module-API/src/main/java/depromeet/api/domain/challenge/controller/ChallengeController.java
@@ -3,14 +3,14 @@ package depromeet.api.domain.challenge.controller;
 import static depromeet.api.util.AuthenticationUtil.getCurrentUserSocialId;
 
 import depromeet.api.domain.challenge.dto.request.CreateChallengeRequest;
+import depromeet.api.domain.challenge.dto.request.JoinChallengeRequest;
 import depromeet.api.domain.challenge.dto.request.UpdateChallengeRequest;
 import depromeet.api.domain.challenge.dto.response.CreateChallengeResponse;
 import depromeet.api.domain.challenge.dto.response.UpdateChallengeResponse;
 import depromeet.api.domain.challenge.usecase.CreateChallengeUseCase;
 import depromeet.api.domain.challenge.usecase.DeleteChallengeUseCase;
+import depromeet.api.domain.challenge.usecase.JoinChallengeUseCase;
 import depromeet.api.domain.challenge.usecase.UpdateChallengeUseCase;
-import depromeet.api.domain.userchallenge.dto.request.CreateUserChallengeRequest;
-import depromeet.api.domain.userchallenge.usecase.CreateUserChallengeUseCase;
 import depromeet.common.response.CommonResponse;
 import depromeet.common.response.Response;
 import depromeet.common.response.ResponseService;
@@ -29,7 +29,7 @@ public class ChallengeController {
     private final CreateChallengeUseCase challengeUseCase;
     private final UpdateChallengeUseCase updateChallengeUseCase;
     private final DeleteChallengeUseCase deleteChallengeUseCase;
-    private final CreateUserChallengeUseCase createUserChallengeUseCase;
+    private final JoinChallengeUseCase createUserChallengeUseCase;
 
     @Operation(summary = "챌린지 생성 API", description = "챌린지를 생성합니다.")
     @ApiResponses(
@@ -81,7 +81,7 @@ public class ChallengeController {
     @PostMapping("/challenge/join/{challengeId}")
     public CommonResponse joinChallenge(
             @PathVariable Long challengeId,
-            @RequestBody @Valid CreateUserChallengeRequest createUserChallengeRequest) {
+            @RequestBody @Valid JoinChallengeRequest createUserChallengeRequest) {
         createUserChallengeUseCase.execute(
                 getCurrentUserSocialId(), createUserChallengeRequest, challengeId);
         return ResponseService.getSuccessResponse();
